+++ zio/logging/src/main/scala/tofu/logging/zio/impl/UIOZLogging.scala
@@ -1,7 +1,8 @@
-package tofu.logging.impl
+package tofu.logging.zio.impl
 
 import org.slf4j.Logger
 import tofu.logging.LoggedValue
+import tofu.logging.impl.LoggingImpl
 import zio.UIO
 
 class UIOZLogging(logger: Logger) extends LoggingImpl[UIO](logger) {
