+++ zio/logging/src/main/scala/tofu/logging/zio/impl/URIOZLoggingImpl.scala
@@ -1,7 +1,8 @@
-package tofu.logging.impl
+package tofu.logging.zio.impl
 
 import tofu.logging.{Loggable, LoggedValue}
 import org.slf4j.{Logger, Marker}
+import tofu.logging.impl.{ContextMarker, LoggingImpl}
 import zio.{UIO, URIO}
 
 class URIOZLoggingImpl[R: Loggable](logger: Logger) extends LoggingImpl[URIO[R, *]](logger) {
